Change: Add `DoesNotMatch` assertion to `DoesNotExtensions_String`
