+++ TUnit.Assertions/Assertions/Strings/DoesNotExtensions_String.cs
@@ -1,6 +1,7 @@
 ﻿#nullable disable
 
 using System.Runtime.CompilerServices;
+using System.Text.RegularExpressions;
 using TUnit.Assertions.AssertConditions;
 using TUnit.Assertions.AssertConditions.Interfaces;
 using TUnit.Assertions.AssertConditions.String;
@@ -56,5 +57,23 @@ public static partial class DoesNotExtensions
             (actual, _, _) => $"\"{actual}\" does end with \"{expected}\"",
             $"not end with {expected}")
             , [doNotPopulateThisValue1, doNotPopulateThisValue2]);
+    }
+    
+    public static InvokableValueAssertionBuilder<string> DoesNotMatch(this IValueSource<string> valueSource, string regex, [CallerArgumentExpression(nameof(regex))] string expression = "")
+    {
+        return DoesNotMatch(valueSource, new Regex(regex), expression);
+    }
+    
+    public static InvokableValueAssertionBuilder<string> DoesNotMatch(this IValueSource<string> valueSource, Regex regex, [CallerArgumentExpression(nameof(regex))] string expression = "")
+    {
+        return valueSource.RegisterAssertion(new FuncValueAssertCondition<string, Regex>(regex,
+                (actual, _, _) =>
+                {
+                    Verify.ArgNotNull(actual);
+                    return !regex.IsMatch(actual);
+                },
+                (actual, _, _) => $"The regex \"{regex}\" matches with \"{actual}\"",
+                $"does not match with {expression}")
+            , [expression]);
     }
 }
